+++ test/setValue.js
@@ -1,5 +1,5 @@
 import test from "ava";
-import FormLinker from "../src/index";
+import FormLinker from "../";
 
 test("set value", t => {
   let fl = new FormLinker({
